Fix: pull 후 발생한 오류 수정
