+++ build.gradle
@@ -36,6 +36,8 @@ dependencies {
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
     testImplementation 'org.springframework.security:spring-security-test'
 
+    implementation 'org.springframework.boot:spring-boot-starter-validation:3.1.5'
+
 }
 
 tasks.named('bootBuildImage') {
