--- src/main/java/com/ll/netmong/domain/post/PostRepository.java
@@ -1,11 +1,0 @@
-package com.ll.netmong.domain.post;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface PostRepository extends JpaRepository<Post, Long> {
-
-    Optional<Post> findById(Long id);
-
-}
