+++ src/main/java/com/ll/netmong/domain/post/entity/Post.java
@@ -1,6 +1,7 @@
 package com.ll.netmong.domain.post.entity;
 
 import com.ll.netmong.common.BaseEntity;
+import com.ll.netmong.domain.postComment.entity.PostComment;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -9,6 +10,8 @@ import lombok.Setter;
 import lombok.experimental.SuperBuilder;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
@@ -20,4 +23,12 @@ public class Post extends BaseEntity {
     @Column(length=100)
     private String content;
     private String imageUrl;
+
+    @OneToMany(mappedBy = "post", cascade = CascadeType.ALL)
+    private List<PostComment> comments = new ArrayList<>();
+
+    public void addComment(PostComment comment) {
+        this.comments.add(comment);
+        comment.setPost(this);
+    }
 }
