+++ src/main/java/com/ll/netmong/domain/post/repository/PostRepository.java
@@ -3,5 +3,10 @@ package com.ll.netmong.domain.post.repository;
 import com.ll.netmong.domain.post.entity.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.Optional;
+
 public interface PostRepository extends JpaRepository<Post, Long> {
+
+    Optional<Post> findById(Long id);
+
 }
