+++ src/main/java/com/ll/netmong/domain/postComment/service/PostCommentServiceImpl.java
@@ -1,12 +1,12 @@
 package com.ll.netmong.domain.postComment.service;
 
 import com.ll.netmong.common.DataNotFoundException;
+import com.ll.netmong.domain.post.entity.Post;
+import com.ll.netmong.domain.post.repository.PostRepository;
 import com.ll.netmong.domain.postComment.dto.request.PostCommentRequest;
 import com.ll.netmong.domain.postComment.entity.PostComment;
 import com.ll.netmong.domain.postComment.repository.PostCommentRepository;
 
-import com.ll.netmong.domain.post.Post;
-import com.ll.netmong.domain.post.PostRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
